Refactored to delete obsolete and depreciated code Renamed Activities etc for consistency
