+++ JImv2/app/src/main/java/com/example/jimv2/CalendarActivity.java
@@ -46,7 +46,7 @@ public class CalendarActivity extends AppCompatActivity {
     public void backToLanding(){
         Log.d(TAG, "backToLanding: called.");
 
-        Intent intent = new Intent(this,landing.class);
+        Intent intent = new Intent(this, LandingActivity.class);
         startActivity(intent);
     }
     public void openPlanRoutine(){
