+++ JImv2/app/src/main/java/com/example/jimv2/LoginActivity.java
@@ -7,15 +7,12 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-public class MainActivity extends AppCompatActivity {
+public class LoginActivity extends AppCompatActivity {
 
     private Button button;
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.login);
         button = (Button) findViewById(R.id.login_button);
         button.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -25,7 +22,7 @@ public class MainActivity extends AppCompatActivity {
         });
     }
     public void openLandingScreen(){
-        Intent intent = new Intent(this,landing.class);
+        Intent intent = new Intent(this, LandingActivity.class);
         startActivity(intent);
     }
