--- JImv2/app/src/main/res/layout/landing3.xml
@@ -1,185 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:background="@color/white"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-
-    <Button
-        android:layout_width="175dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginStart="20dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginTop="25dp"
-        android:layout_marginEnd="217dp"
-        android:layout_marginRight="217dp"
-        android:background="@drawable/rectangle"
-        android:baselineAligned="false"
-        android:clickable="true"
-        android:text="Rank"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textSize="32dp" />
-
-    <ImageButton
-        android:id="@+id/leftArrow"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_centerVertical="true"
-        android:layout_marginStart="-3dp"
-        android:layout_marginLeft="-3dp"
-        android:layout_marginTop="292dp"
-        android:background="@android:color/transparent"
-        android:clickable="true"
-        app:srcCompat="@mipmap/left2_arrow" />
-
-    <TextView
-        android:id="@+id/exercise_list"
-        android:layout_width="291dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerInParent="true"
-        android:layout_marginTop="162dp"
-        android:height="280dp"
-        android:background="@drawable/rectangle"
-        android:textColor="@color/white"
-        android:textSize="40dp" />
-
-    <!--    <Button-->
-    <!--        android:id="@+id/quickAddExerciseButton"-->
-    <!--        android:layout_width="200dp"-->
-    <!--        android:layout_height="40dp"-->
-    <!--        android:layout_centerHorizontal="true"-->
-    <!--        android:layout_marginTop="390dp"-->
-    <!--        android:background="@android:color/transparent"-->
-    <!--        android:gravity="center"-->
-    <!--        android:text="(+)"-->
-    <!--        android:textColor="@color/white"-->
-    <!--        android:textSize="20sp" />-->
-
-
-
-    <Button
-        android:id="@+id/workoutButtonLanding"
-        android:layout_width="140dp"
-        android:layout_height="60dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="34dp"
-        android:layout_marginLeft="34dp"
-        android:layout_marginBottom="92dp"
-        android:background="@drawable/rectangledb"
-        android:gravity="center"
-        android:text="Workout"
-        android:textColor="@color/white"
-        android:textSize="20sp" />
-
-    <Button
-        android:id="@+id/calendar_button"
-        android:layout_width="140dp"
-        android:layout_height="60dp"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginEnd="38dp"
-        android:layout_marginRight="38dp"
-        android:layout_marginBottom="92dp"
-        android:background="@drawable/rectangledb"
-        android:gravity="center"
-        android:text="Calendar"
-        android:textColor="@color/white"
-        android:textSize="20sp" />
-
-
-    <Button
-        android:id="@+id/friends_button"
-        android:layout_width="140dp"
-        android:layout_height="60dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginStart="35dp"
-        android:layout_marginLeft="35dp"
-        android:layout_marginBottom="21dp"
-        android:background="@drawable/rectangledb"
-        android:gravity="center"
-        android:text="Statistics"
-        android:textColor="@color/white"
-        android:textSize="20sp" />
-
-    <Button
-        android:id="@+id/statistics_button"
-        android:layout_width="140dp"
-        android:layout_height="60dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginStart="233dp"
-        android:layout_marginLeft="233dp"
-        android:layout_marginBottom="23dp"
-        android:background="@drawable/rectangledb"
-        android:gravity="center"
-        android:text="Friends"
-        android:textColor="@color/white"
-        android:textSize="20sp" />
-
-    <TextView
-        android:id="@+id/currentDate"
-        android:layout_width="wrap_content"
-        android:layout_height="35dp"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="120dp"
-        android:background="@drawable/rectangledb"
-        android:clickable="true"
-        android:ems="10"
-        android:inputType="date"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textSize="20dp" />
-
-    <ImageButton
-        android:id="@+id/userIcon"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginStart="250dp"
-        android:layout_marginLeft="250dp"
-        android:layout_marginTop="13dp"
-        android:layout_marginEnd="17dp"
-        android:layout_marginRight="17dp"
-        android:clickable="true"
-        android:background="@android:color/transparent"
-        app:srcCompat="@drawable/shrek_icon2" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/landingRecyclerView"
-        android:layout_width="285dp"
-        android:layout_height="281dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_centerInParent="true"
-        android:layout_marginStart="63dp"
-        android:layout_marginLeft="63dp"
-        android:layout_marginTop="162dp" />
-
-</RelativeLayout>
