+++ JImv2/app/src/main/res/layout/login.xml
@@ -5,7 +5,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".MainActivity">
+tools:context=".LoginActivity">
 
     <Button
         android:id="@+id/login_button"
